In order to represent the changes more accurately, I hacked the grammar a smidgeon. Instead of replacing an AST_if with another AST_if, only with statements lists each containing a single goto, I added an AST_hir_if, whose iftrue and iffalse field contain a non-null goto statement.
The problem which necessitated this was that I generated ifs, and these ifs were later transformed by another lowering. This happened, for example, with nested loops. The inner loop was transformed first creating an if statement, and the outer (for) loop was then transformed by creating a while, and transforming it. This, however, transformed the inner loop, to create code which wasnt wrong, but was very difficult to understand. So after this transform, there are no AST_ifs, only AST_hir_ifs.

I also added unparsers for AST_hir_if, and some test cases, nearly all of which work. I renamed 003, since its just a for-loop test.


git-svn-id: http://phc.googlecode.com/svn/trunk@196 bbcf20a9-8526-0410-90e2-9b3d3f589b1d