Added an interpret_lowered test, and added one bugfix to the goto_uppering pass (the list of array indices of a variable cannot be NULL, although of course it can be empty), but the goto_uppering doesn't seem to be completely right; when I run it on the factorial test, for example, it doesn't add a switch case for the "start" label. I'll leave that for Paul to fix though :)
Also, we should apply goto_uppering to the list of statements in every method
(which it does), and also the "global" list of statements (which it doesn't at
the oment). I think we can probably just apply the exact same code to both
lists without it breaking (you'd get funciton definitions inside switch
statements, but who cares?). The reason is that we can then test lowering and
shredding without running lifting -- the problem with lifting is that the code
generator will have to treat __MAIN__ special (it's local variables aren't
really local variables, but global variables), so it cannot trivially be tested
(there probably are ways, but it'd be nice to make the interpretshredded and
interpretlowered tests independent of lifting).



git-svn-id: http://phc.googlecode.com/svn/trunk@222 bbcf20a9-8526-0410-90e2-9b3d3f589b1d