Ok, it turns out that "print" isn't really a function either, so the translation from "echo" to "print" doesn't help. That means that we have to translate both echo(x) and print(x) to printf("%s",x). It is necessary to use the "%s" construct to avoid printf interpreting the string.
Incidentally, the code that does this illustrates a nice new feature of the
matching API. We set up two patterns, but they share a wildcard:

    Wildcard<AST_expr>* arg = new Wildcard<AST_expr>;
    AST_method_invocation* echo = new AST_method_invocation(
        new Token_class_name(new String("%STDLIB%")),
        new Token_method_name(new String("echo")),
        new List<AST_actual_parameter*>(
            new AST_actual_parameter(false, arg)
        ));
    AST_method_invocation* print = new AST_method_invocation(
        new Token_class_name(new String("%STDLIB%")),
        new Token_method_name(new String("print")),
        new List<AST_actual_parameter*>(
            new AST_actual_parameter(false, arg)
        ));

Then to use the wildcard, we simply do

    if(in->match(echo) || in->match(print))
    {
		// use arg->value here
	}

So two advantages: first off, it is really easy to access the value of the
wildcard (previously, that would have been more awkward because we never had a
handle on the wildcard itself). Secondly, because the wildcard is shared by
both patterns, inside the if we do not need to know which one of the two
patterns matched. 


git-svn-id: http://phc.googlecode.com/svn/trunk@244 bbcf20a9-8526-0410-90e2-9b3d3f589b1d