Regenerated test files, and added new ones. Added a few more attribute cases to XML_unparser.
Stripped commented code from Process_includes. This is going to need to be reworked. It barely works as is (hence i didnt update test files for that). I think the best idea is to try and include in every case where we know the filename, whereas now we only include files included with include () and require () with a string constant argument, and no return value.



git-svn-id: http://phc.googlecode.com/svn/trunk@277 bbcf20a9-8526-0410-90e2-9b3d3f589b1d