Finally got tired of PHPs changing behaviour. It turns out we dont properly support octal. This is a problem with strtod, which doesnt like it. PHP changed this in 5.2.2, which also changed how the edge cases of ints specified as hex, turn to floats. It made PHP more consistent, but it means we had to change our code. Not only that, but we had to write/steal an octal stdtod function (which is what PHP does). Since we will later need the embedded interpreter (say, for constant folding), I tried it now. Works well. We call an eval, and strip out its type and value.
However, this makes the code very slow. I'd estimate twice as slow for the tests. This is almost certainly because I start and stop the interpreter every time. I'll need to stop doing that, but I couldnt get it to compile with the "start the interpreter in one function and end it in another" way presented in the PHP internals book, but I'll try again.

All the embedded stuff is safely abstracted in embed/, using static functions of the PHP class. I think its wise to keep it from spreading.

The autotools stuff is probably not perfect, but works for now.

Added a negative literals test
Include php path in LD_LIBRARY_PATH when running phc in the tests
A number of regressions in the AST had the wrong type - edge cases we hadnt spotted. Regenerated.
make generate-test-files generated long ones, which now takes forever. Added long-generate-test-files and changed the other to only do short tests.



git-svn-id: http://phc.googlecode.com/svn/trunk@306 bbcf20a9-8526-0410-90e2-9b3d3f589b1d