A number of changes to errors.
# phc_warning / phc_error
There were two problems with warnings. Firstly, they're too hard to add. They must be added in 3 different files, one of which (error.h) forces the recompile of the entire project. In addition, they're not local, so changing them one involves looking up another file.

Now, instead, the warning/error is called by phc_warning (format_string, filename, linenumber, ...). I also added a convenience function phc_warning (format_string, ...), which surprised me a little bit by working.

Changed all the call sites, and places which had multiple uses of the same variable were switched to a local #define.


# deja-gnu style warnings
Issuing compile time warnings was tough to handle in the testing framework. They have to be marked as erroneous, and generate support files. In addition, phc's stderr was frequently passed on to the php interpreter, which was obviously wrong. As a result, a lot of them were moved to unsupported.

Instead, if a warning or error is expected, we modify the test files as follows
<?php
	$x[]; // { Error: Cannot use [] for reading }
?>
where the {} contains the exact error issued by phc, with the line number and filename stripped. We only support 1 error per file, obviously.

Since we now have test dependencies, we no longer need erroneous with this construct. I havent removed it yet, but will soon.


# Invalid PHP checking pass
The pass Invalid_check issues errors in the presence of constructs that PHP doesnt like. For now, it only finds $[] constructs. Its run after parsing, but could be run anytime.




git-svn-id: http://phc.googlecode.com/svn/trunk@355 bbcf20a9-8526-0410-90e2-9b3d3f589b1d