Theres been a couple of bugs with the Shredded and arrays, so we decided its best to leave them as is for codegen. The most prevalent problem was an assignment of an array to a class attribute in an initializer. This would result in the members of the array being 'push_back'ed, and often dissappearing completely. Another problem was empty arrays conferred type information, so warnings such as "each() not passed an array" would appear.
We could fix the latter with a cast, though it would be a little hacky. The first problem though is mostly insurmountable. Edsko suggested we add a special constructor that did the initialization. However, knowing when to call that would be difficult. In particulr, if a subclass overrode a parent's constructor, we wouldnt know when to initialize the parent's variables. As a result, its best to handle the problem in code generation.


git-svn-id: http://phc.googlecode.com/svn/trunk@380 bbcf20a9-8526-0410-90e2-9b3d3f589b1d