Add some tests.
The first is an exmapel of why we couldnt add the special constructor to shred arrays.
The second shows some pleasant features of run-time class declaration.
The third uses an illegal offset for a key, and produces an error. No check is in place for this yet.
The last one uses a function in array declaration


git-svn-id: http://phc.googlecode.com/svn/trunk@384 bbcf20a9-8526-0410-90e2-9b3d3f589b1d