This may break the build!!
When PHP wasnt found by configure, nothing worked. Made a number of changes to the autotools, and to the tests to make it work.

The prerequisites now checks for the $php variable, instead of the autovars $php_exe. If no autovars is found, the PATH is searched (using which) to find a php to run the tests with.

If autoconf doesnt find an embed sapi, make will still run the test if it finds php in the path, again using which.

phc wouldnt compile without the embed SAPI, due to changes I made before while embedding it. The first problem was that the build assumed the presence of the SAPI, which has been overcome by making it use the autoconf machinery to only conditionally add the files. The second problem was that there was no #ifdef test for the preprocessor, which has been added, and that there was no HAVE_PHP macro, which has been added. We now check for the macro, and in its absense, use the old Token_conversion check, which worked as of 5.2.1, but wont properly work now.

More tests are broken than should be, but it compiles without PHP. I still have to check that it compiles properly with PHP, and to fix the tests.



git-svn-id: http://phc.googlecode.com/svn/trunk@407 bbcf20a9-8526-0410-90e2-9b3d3f589b1d