Finally fix the nasty braces problem. The problem has about 4 parts to it. Firstly, we rename needs_curlies to needs_user_curlies, to indicate these came from the user. We rename needs_brackets to needs_user_brackets for consistency.
We then identify the cases where curlies are required, and add them in the unparser. Curlies are required based on context, so only specific parent/child combinations need them, with the child always being an AST_reflection. So we add the needs_curlies attribute in the parent in those cases, and print a curly in the event of either needs_curlies or needs_user_curlies. This adds attributes, which is _slightly_ wrong. I suppose they could be removed on the way out if required.

We update the canonical unparser, by removing all user curlies and brackets. We clone the tree first to ensure we dont make a mistake here.

The AST_expr constructor added the needs_brackets attribute. This is pointless, so I've stripped it.

Finally, I removed the dependency of InterpretCanonical on InterpretUnparsed. I'm not sure if this is right, but InterpretCanonical might work when InterpretUnparsed doesnt.


It seems my previous commit failed, so I have an additional commit here.

I amalgamated all off the checking functions into Consistency_check, and they can be checked at anytime by calling check ().



git-svn-id: http://phc.googlecode.com/svn/trunk@427 bbcf20a9-8526-0410-90e2-9b3d3f589b1d