Split the binops example into 4 smaller examples, and fix them.
The problem was that
	$x = array (true, false);

was being shredded into
	$x[] = true;
	$x[] = false;

so that when this was done in a loop, you got a larger array then intended. The solution is to add
	unset ($x);
	$x = (array) $x;

first, so that the var is cleared and the type is set. Conveniently, this is also a way to shred
	$x = array ();

which is now no longer in the HIR.


git-svn-id: http://phc.googlecode.com/svn/trunk@435 bbcf20a9-8526-0410-90e2-9b3d3f589b1d