I was using these a while back ,and they're kinda useful, so I may as well commit them.
If you run:

	misc/phcdiff --run-shredder --run-goto-uppering a.php

it will run diff the output of php on the original file and php on the shredded file. You can use any options, as long as the filename goes last.

xmldiff shreds both and diffs the output.

If we break these in the future its no harm, but the'yer a smidgeon useful.


git-svn-id: http://phc.googlecode.com/svn/trunk@442 bbcf20a9-8526-0410-90e2-9b3d3f589b1d