New Pass manager:
Add the pass manager, move everything into passes, and redo the tests and the dumping to work with the pass manager. The pass manager creates a pass queue, and phc.cpp adds all the passes in order. Plugins are then allowed add their passes, and all the passes are then run in order.

A pass extends Pass, has a name and implements the run () method. The run () method gives passes access to the pass_manager, so they can add passes of their own if they wish. We havent used this yet, though.

Every pass is run every time, and its up to the passes run method to do nothing if specified by the command line options. For example, Compile_C does nothing if -c is not specified. This allows us to run most passes all the time, and getting the output of different passes is done with a flexible dumping API.

After each pass is run, the pass_manager will check the command line options to see if this pass should be dumped. There are 4 kinds of dump, which can be invoked after each pass. --dump=x will dump PHP after the pass named 'x'. --udump=x dumps Goto-uppered PHP, -xdump=x dumps XML, and -ddump=x dumps DOT. All passes are named, and a Fake_pass can be added for points in the compilation process that we'd like to name. --dump=ast and --dump=hir are handled this way.

--pretty-print and --obfuscate are a handles by simple passes which run print after checking the command-line options. This is unrelated to the dumping options. Passes from parse(), except parts of the parse process, have moved into the pass queue.


New Plugin API:

Plugins have an API of 3 functions. load () is called after the pass queue is created, and allows the plugin to load its pass anywhere in the pass queue (including multiple times). The Pass_manager has two functions add_after_all_passes () and add_after_named_pass (), which handle all the use cases I had. Plugins are implicitly wrapped in a Plugin_pass, which is passed to load (), so plugins dont have to worry about creating this.

The run () function is called when its that passes turn to run. Finally, the unload function is called once per pass, even if the pass is in multiple places in the queue. We use this to print "success" in a number of tests.



Some caveats and other notes:

--dump=tokens doesnt work. I'll have to add this specially.

I removed the regression_dump_includes test. Includes is going to be always turned on, so we disable this. I also removed the compile-time-includes option.

Note that the tutorials have not been updated. We really should add test cases for these, so we can test that they work. I presume they will segfault if they are called.

InterpretLowered didnt work with --dump=lcf, only with --dump=lef. Added new test for this. Will fix soon.

cloning, linear and source_vs_semantics and pre_vs_post_count now run after every pass. No bugs found.

I cant get automake1.10 to work, so 1.9 is used. I'll regenerate with 1.10 when I get back.



git-svn-id: http://phc.googlecode.com/svn/trunk@447 bbcf20a9-8526-0410-90e2-9b3d3f589b1d