Due to the previous problem with 	$x[f()] += 1; I modified canonical_unparser to disregard the is_opeq attribute, which leads opeq.php and long_opeq.php to fail, which is correct.
This led to an error involving a conditional expression, so I modified the canonical_unparser to correctly bracket everything in its path, which fixed the problem.

However, this led to some investigation of what's legal as an AST_directive, and I added a test case showing "ticks" in case-insensitive. Since we dont have a case-insensitive string comparison function (its done in the lexer using builtin lexer features), I added compare_ci to the String class. Also, since ticks is the only allowed directive, added an error if another directive is specified. Perhaps this should be a warning, however. A test case (sticks.php) is also added for this.

Finally, the same problems which affect directives also affect attributes, static_declarations and formal_parameters, so Invalid_check was extended to give the same warnings for all (as before, some errors are enforced by the parser, so tests wernt added).



git-svn-id: http://phc.googlecode.com/svn/trunk@453 bbcf20a9-8526-0410-90e2-9b3d3f589b1d