Add some tests cases I hadnt though of before, including one for a bad type hint, and a break with an expression which echoes.
Generate_c should be called regardless of the --generate-c flag. It should, however, not print unless the flag is set.

Print a proper error message for multiple break levels. I really should have homogenized the output instead, but I forgot I had done that. Its still not proper, since it says "1 levels" instead of "1 level", but that can be fixed later.


Note this isnt fully tested since the machine I'm working on is going down in a minute.




git-svn-id: http://phc.googlecode.com/svn/trunk@463 bbcf20a9-8526-0410-90e2-9b3d3f589b1d