To emulate php_cli, we set display_startup_errors and report_zend_debug (the later is only used for printing errors which would be double printed if we didnt turn it off).
In order to provide filenames and line numbers in error messages, we set some of the interpreter's fields, wrapping them with phc_setup_error, so that they'll work if the error doesnt cause the program to fail.

This provides error messages just like php cli (though this isnt used in all places, its good enough for now - especially as tests pass).


git-svn-id: http://phc.googlecode.com/svn/trunk@493 bbcf20a9-8526-0410-90e2-9b3d3f589b1d