Some small changes to make the refocunt test work.
I was printing out the HIR, so labels and gotos hadnt been removed yet. Any tests with flow control failed as a result. Now print out the uppered hir, after the test, using the --udump= mechanism.

Also, if we make the test dependent on CompliedVsInterpreted, there's no need to have a generate-c pass. This slightly defeats the purpose of using this test to fix the CompiledVsInterpreted, however. I'll fix this in a later commit.


git-svn-id: http://phc.googlecode.com/svn/trunk@494 bbcf20a9-8526-0410-90e2-9b3d3f589b1d