Added the split test from the last commit. null_obj_ref leads to an error that was removed from horrible/obfuscated_foreach.php.
I suspected there was a problem with lowering of switch statements, from the comment in Lower_control_flow.cpp. Added switch_side_effects.php, which is designed to check that the switch expression isnt evaluated more than once. It turns out it was the comment that was wrong, so I fixed it.

In a foreach statement of the form "foreach ($arr as $key => $value)", $key should only be written to in the loop body, and should retain the final key value, not NULL, after the loop ends. We changed it to NULL. Fixed.


git-svn-id: http://phc.googlecode.com/svn/trunk@519 bbcf20a9-8526-0410-90e2-9b3d3f589b1d