This adds two tests, each of which convert half the statements to 'eval ($unparsed_stmt);' and execute them like this. The tests also include the debug_zval_dump, though I think this is probably a mistake, since the tests take more than half an hour to run.
The two tests use the same plugin, but provide a command line arguemnt to it. It was therefore necessary to add the mechanics to support this. This involved adding a command like option '-r-option', in the same vein as '-c-option'. -r-option[i] is assocaited with the i_th plugin provided.


git-svn-id: http://phc.googlecode.com/svn/trunk@571 bbcf20a9-8526-0410-90e2-9b3d3f589b1d