Add an optimize option to phc. It essentially chooses between "-O3 -g -NDEBUG" and "-O0 -ggdb3".
misc/comp now respects the flag, and links to an optimized libphp5.so.

With no optmizations, the ratio of PHP/phc is 29s/177s or so (about 6x slower). Now the ratio is 7s/46s, which is about 6.5 times slower.

With optimization, about 42% of the time is spent is zend_hash_quick_find or zend_hash_quick_add_or_update, and 28% of the time in memory allocation.


git-svn-id: http://phc.googlecode.com/svn/trunk@662 bbcf20a9-8526-0410-90e2-9b3d3f589b1d