Add constant pooling, predicated on the -O flag. Added a new test for it aswell. 5 test subjects fail this, but I'm not sure why.
If -O is not specified, code for literals is generated as before. If it is specified, a literal assignment will generate an assignment to a variable such as 'phc_const_pool_int_0'. The declaration, definition and finalization (cleanup memory) of this variable will also be generated. We could probably have 1 variable less here, as NULL could just use EG (uninitialized_zval_ptr), but I think this is fine.

This runs about 15% faster than without the pooling.


git-svn-id: http://phc.googlecode.com/svn/trunk@674 bbcf20a9-8526-0410-90e2-9b3d3f589b1d