Part 3 of possibly 4 parts of the st_entry_not_required_patch.
In order to get correct behaviour for GLOBALS, we must have a local_GLOBALS available. Instead of hacking it, it is clearer to make this explicit in the HIR used to generate code.

Clarify adds 'global $GLOBALS' at the start of every function. We do this after what we technically call HIR, so that its available for codegen. This removes the global ("GLOBALS) hack we used to have in method_entry. It also means that superglobals are available for pruning. As a result, we move the pruning stage to after this, but before codegen, which is a more natural place for it anyway.


git-svn-id: http://phc.googlecode.com/svn/trunk@713 bbcf20a9-8526-0410-90e2-9b3d3f589b1d