The final commit of the st_entry_not_required series.
The majority of this commit is changing functions to use Token_variable_names instead of AST_variables. This involves adding a lot of get_var_name () calls.


We also remove the call to global ("GLOBALS"), since that is now explicit in the HIR, due to the Clarify pass. The global(char*) function is therefore no longer required.

generate_non_st_declarations() and generate_non_st_cleanups now search in formal_parameters, since we intend for them to use local variables instead of symbol table entries as well.

The final part is to add a bit of hacky code to deal with params. It seems that this should use a combination of read/write/separate just like global () does, but this seems to work.


Overall, we've shaved the running time from 28.5 seconds to 22 seconds, meaning we are roughly 3 times slower than the interpreter now.


git-svn-id: http://phc.googlecode.com/svn/trunk@715 bbcf20a9-8526-0410-90e2-9b3d3f589b1d