Remove some unnecesarry steps from ht_* functions.
We dont need to allocate a new zval for double conversions, a C level cast is fine.

We dont need to make a copy for string indexing, since its already a string.

We dont need a zend_symtable_* call for NULL, just a zend_hash_* call is fine.

Some indentation and spellings and TODO also included.


git-svn-id: http://phc.googlecode.com/svn/trunk@730 bbcf20a9-8526-0410-90e2-9b3d3f589b1d