Implement better caching of function calls. A lot of our profiling time was spent in call_user_ex, since call_user_ex doesnt as the function pointer as one of its handlers.
The solution is built into the Zend API, but not documented in Sara's book. We create a zend_fcall_info_cache, as well as a zend_fcall_info. We do this using 'static'. Doing it the other way - an ahead of time pass - would be slightly faster, as it would run once, as opposed to once per source code occurence. This isnt too bad though.

The are a number of tricky bits. The first is that no_separation is turned on by zend_fcall_info_init. We seem to separate correctly, but 1 change was required to fetch_array_arg_by_ref, where we seem to get it wrong.

Secondly, its all fine and good to put your parameters and return values in the zend_fcall_info structure, but we want to avoid overwriting them with recursive calls. So we save the previous contents in the local stack frame. Another alternative exists, in which we use zend_fcall_info_call, but that requires the paramters to be in a hashtable, so this works better.


This results in a 16% speed improvement, finally making us faster than the PHP interpreter.


git-svn-id: http://phc.googlecode.com/svn/trunk@746 bbcf20a9-8526-0410-90e2-9b3d3f589b1d