Rathan than convert pre- and post-op in the following manner:
	$x++;
becomes
	$x = $x + 1;

we leave it as is. We determined that we need to use the increment_function(), in order to handle strings properly, so we must keep this.

At the same time, an effort is made to reduce the amount of generated code. As opposed to 4 extra statement which were created as a result of the particular interactions of passes, we now only generate 1 1/2 statements (I'm considering a half-statement to be where we dont write back to the LHS) for post-op, and only 1/2 for pre-op.

We convert these so that whether its a pre-op or post-op doesnt matter, and we discard the LHS for this purpose.

As a result of reducing the amount of generated code, as well as using increment function instead of add_function, this lowers the run-time by another 7%, so we are now 25% faster than PHP.


git-svn-id: http://phc.googlecode.com/svn/trunk@760 bbcf20a9-8526-0410-90e2-9b3d3f589b1d