This commit paves the way for introducing the HIR bit-by-bit. I have duplicated the AST definition in phc.tea as the new "HIR" definition in hir.tea; that is, the current "HIR" definition is exactly identical to the AST definition with the exception of the prefix and the namespace used. The most significant diff in this log is probably:
-class Generate_C : public AST_visitor, public Pass
+class Generate_C : public HIR_visitor, public Pass

In other words, Generate_C is no longer an AST visitor, but an HIR visitor. I
have defined a (trivial) translation from the AST to the HIR in AST_to_HIR.h,
which uses the new Fold feature of maketea (so, this requires maketea revision
57 or higher). The general idea is that we can now start modifying the HIR
iteratively, making the translation from the AST to the HIR slightly more
complicated as we go, and at the same time simplifying the code generation. 

To faciliate disambiguation, I have moved all the AST stuff to a namespace of
its own ("AST"). Many of the changed files in this commit thus only contain a
single modification "using namespace AST".


git-svn-id: http://phc.googlecode.com/svn/trunk@779 bbcf20a9-8526-0410-90e2-9b3d3f589b1d