Add a --include option (doesnt do anything just yet).
Add a --list-passes option, which lists passes' names, descriptions, and whether they are enabled.

Added an pass_is_enabled method to Pass, for a pass to predicate if it is enabled, based on the Pass_manager's settings (ie command-line arguments). This isnt widely used, but I've implemented it from --pretty-print, and it works.

I'm going to move other checks to use this, rather than checking inside the run function.


To remove a pass, simple call the (new) get_pass function on the pass manager, and call set_enabled on it.


git-svn-id: http://phc.googlecode.com/svn/trunk@790 bbcf20a9-8526-0410-90e2-9b3d3f589b1d