This is a first but important step towards supporting in-string syntax in the unparser. In this commit, the lexer marks concatenation operators as "phc.unparser.in_string_syntax.simple" (but only for the simplest case, "foo $a bar"). This attribute is taken into account in the unparser. This is a more difficult problem that it may seem due to the effectively arbitrary associativity of the concatenation operator. The solution is to linearize the tree, collecting all concat arguments, and then partitioning this list into sublists which form strings. This is done by a visitor "Linearize", defined within the AST_unparser.
For a full solution, the lexer needs to mark *all* concatenation operators that
we insert, we need to be able to copy the style of syntax the user is using
(normal $a, delimited ${a}, or complex {$a}), and finally we need to be able to
unparse HEREDOC strings. However, with this infrastructure now in place this
shouldn't prove too difficult.


git-svn-id: http://phc.googlecode.com/svn/trunk@838 bbcf20a9-8526-0410-90e2-9b3d3f589b1d