Fix a number of problems related to casts. Initially, this was an attempt to unparse the source_rep for casts, so that "(bool)" would print as "(bool)", as opposed to "(boolean)".
This began by capturing the source_rep. The field was added
in phc.tea, the lexer now stores the source_rep, and the
parser now passes it. As part of the lexer fix, String got a
new kind of constructor. 

Fixing the lexer also allowed me to
accept a slightly smaller subset of whitespace, since PHP
doesnt allow newlines in the definition. bad_cast.php tests this.

I also allowed mixed case in the typename, and added a test to casts.php for this.

Finally, we use an array of scalars in a few places, so best to extract them into an included file, scalar_array.php.


git-svn-id: http://phc.googlecode.com/svn/trunk@841 bbcf20a9-8526-0410-90e2-9b3d3f589b1d