Unfortunately, I've done too much work to put in one commit, so I've split them up. This should slightly break the build, but not a huge amount. It will be fixed in a moment.
This abstracts the XML_unparser so that it can be used as both an AST_XML_unparser and a HIR_XML_unparser.

I've removed as many #includes of XML_unparser as I can for now. It takes a while to compile, since all the work is in the header. I've moved debugging functions into debug.h as a result.

Unfortunately, this breaks the xml_roundtrip test. This should be fixed momentarily.


git-svn-id: http://phc.googlecode.com/svn/trunk@861 bbcf20a9-8526-0410-90e2-9b3d3f589b1d