This should fix the build.
Begin the move to using the new HIR. Note that this breaks a lot of CompiledVsInterpreted tests, but better to commit now, and fix them later.

This replaces a lot of exprs with VARIABLE_NAMEs.

It also changes the namespacing of HIR_unparser, since we want to use it more widely.


git-svn-id: http://phc.googlecode.com/svn/trunk@863 bbcf20a9-8526-0410-90e2-9b3d3f589b1d