The idea of this pass was to make the AST fold directly into the identical HIR. This would allow me to move a single pass into the AST queue at a time, and minimize the size of future revisions. In practice, this touched a huge number of files, and so this revision is much larger than I'd hoped. It also didnt fix everything, so a number of features are just broken.
Some passes were left in the AST_to_HIR directory, since they definitely belong there. AST_lower_expr was added to handle the AST_to_HIR lowering (automake doesnt like two files with the same name, even if they are in different directories).

Since the AST and the HIR are the same, the fold had to be expanded to support all of the constructs. Hwoever, some are going to be deleted just as quickly as they were added, but not for a little while.

Some plugins still dont work, that's OK. They'll be fixed soon. The work to fix them would just be undone later, and they arent a priority.

Likewise, Process_includes no longer works on the HIR. TODO.

To debug some problems, I tried to use the reduce script. However, it really relies on compiling, so it didnt help with the problems. However, it now has a -f flag to provide the original XML.

Since there are multiple node types, we add a few phc_error functions to handle them in the same way as AST::Node. However, this leads to a lot of code duplication, and templating doesnt work properly (due to the presence of the functions with var_args), so I meta-templated error.cpp using macros, to avoid duplication.


git-svn-id: http://phc.googlecode.com/svn/trunk@943 bbcf20a9-8526-0410-90e2-9b3d3f589b1d