Switch statements can't be lowered in the same way that of control-flow statements can. In particular, if you have an expression in a switch_case, where is it going to go? So if we leave switches in the HIR, we have to leave complex expressions in the switch_case expressions.
Instead, we lower them in the AST, and remove them from the HIR. They are reduced to if statements in the AST, with a do_while loop around them to account for breaks and continues.

To avoid a crash, we also return "<unknown>" for get_filename, instead of NULL. This is a little bit of a hack, but I dont want to spend a lot of time following filenames around until we're done with larger things.


git-svn-id: http://phc.googlecode.com/svn/trunk@1047 bbcf20a9-8526-0410-90e2-9b3d3f589b1d