The plugins' signatures were not checked at any point, due to C style linking. The compiler was convinced by the type cast that the types were correct, but this created a lot of strange and unpredictible bugs as a result.
This adds a set of type signatures for plugins, in Plugin_pass.h, which all plugins must include anyway. It also fixes all the incorrect signature problems, of which there were a lot. This doesn't fix Refcount and Demi_eval tests, but it moves them on to the next problem.



git-svn-id: http://phc.googlecode.com/svn/trunk@1115 bbcf20a9-8526-0410-90e2-9b3d3f589b1d