Currently, to unparse the Foreach_* constructs, the AST_unparser generates ArrayIterator things. This approach is a little ugly.
A better approach is that we upper the constructs in the MIR. This means that when we --dump the MIR (no uppering), we get a real representation, rather than the uppered version.

This also makes each Foreach_* construct take a VARIABLE_NAME instead of a variable, which makes it more consistent through IRs.


git-svn-id: http://phc.googlecode.com/svn/trunk@1119 bbcf20a9-8526-0410-90e2-9b3d3f589b1d