This adds the parse_buffer, a simple way of abstracting away all the boiler-plate code for creating nodes.
The very old way of creating nodes was:

	out->push_back (new Eval_expr 
		(new Assignment (
			result->clone (),
			NULL,
			new Bin_op (
				temp1->clone (),
				new String ("+"),
				temp2->clone ()))));

The slightly newer way of doing it is (this varies depending on when I used it:

	stringstream ss;
	ss << "$" << *(result->get_value_as_string) << " = $" << *(temp2->get_value_as_string) << " + " << *(temp2->get_value_as_string) << ";"
	out->push_back_all (lower_ast (new String ("tidyp"), parse_to_ast (new AST_unparser (new String (ss.str()), true), in)));

The latest way of doing this is to use the new Parse_buffer and the stream operators:
	(*out
		<< "$" << result << " = $" << temp1 << " + $" << temp2 ";"
	).to_pass (s("tidy_p"), in);

which is cleaner and more obvious, if not hugely shorter.

Note that neither of the parsing approaches does a complete job, since clone copies the attributes aswell. I think this can be accomplished using the Clone_blank_mixins, but I'll need to work on this in the future.


git-svn-id: http://phc.googlecode.com/svn/trunk@1125 bbcf20a9-8526-0410-90e2-9b3d3f589b1d