In investigating the strcat slowdown, it seems that a lot of assignment overwrite themselves. Variables which have a ref-count of one will decrement their refcount, then overwite themselves with the new value, then increase the ref-count. We add a check to see if the variables are the same before this.
Note that we check outside of write_var, as opposed to moving the check into write_var. Initially I did the latter, but only got a small speed improvement. It seemed that since this was happening in a loop, the branch would go the same way each time. But if there is only 1 branch location, inside write_var, then the predictor will get confused. If it is outside it, then each condition will get its own predictor, which should be nearly 100% predictable. Sure enough, this got another speedup.

Overall, this represents a 4-5% speedup.


git-svn-id: http://phc.googlecode.com/svn/trunk@1195 bbcf20a9-8526-0410-90e2-9b3d3f589b1d