There is some confusion between AsyncTest::mark_failure and Test::mark_failure. Rename AsyncTest::mark_* to async_*, and add necessary functions. This obviously changes the signature for mark_success.
I'm not sure why out_handlers weren't bailing out on error, but it makes more sense that way. Combined with the other changes, this also allows us to replace some finish() functions with a simple call to async_success.

Convert ParseASTDot to use AsyncTest. This shortens the code and makes it more consistent.




git-svn-id: http://phc.googlecode.com/svn/trunk@1218 bbcf20a9-8526-0410-90e2-9b3d3f589b1d