Move Foreach_uppering into the MIR, where it really belongs. This is the first step in removing MIR constructs from the AST. The only complication is that now the statements from the parse_buffer are shredded. This leads to statements like: 	$x =& $T = $key;
which apart from not being allowed in the MIR, are not parsable. So we extract the assignment from the assignment, and return that instead, leading to:

	$x =& $key;

which is correct.

This also adds a test, reduced from a bigger test. However, this threw an exception (since we use ArrayIterators for Foreach_uppering), instead of a warning. Add a check that the variable is an array, and issue a warning if it is not.


git-svn-id: http://phc.googlecode.com/svn/trunk@1234 bbcf20a9-8526-0410-90e2-9b3d3f589b1d