This removes Foreach_* constructs from the HIR, and lays the ground for having purer IRs (ie removing label, branch, etc).
In lower_control_flow.cpp, we now wrap each construct in a foreign_expr or a foreign_statement. This means that when they get folded, it is simply enough to extract them from thei Foreign_* container, and there is no need to fold them.

It is not possible to just have a single Foreign node, we instead require a Foreign_expr and a Foreign_statement. We might potentially require more in future if we have nodes in the MIR which would be stored as other HIR nodes (see comment in ast.tea for the implications).


git-svn-id: http://phc.googlecode.com/svn/trunk@1254 bbcf20a9-8526-0410-90e2-9b3d3f589b1d