Remove Label, Branch, LABEL_NAME and Goto from the AST. The hardest part here was fixing the Goto_uppering pass. Most of the other changes were mechanical.
The Goto_uppering creates a large AST structure within the MIR, which took a little fiddling with to get correct. The main subtlety is that putting MIR::Statements in AST::Foreign_statement nodes is a mistake. Its fine for MIR::Foreach_* etc, as these will be unparsed by the MIR_unparser. However, an MIR::Eval_expr, for example, won't be, as the MIR_unparser will call visit() on this, for which there is no handler (to unparse it, it would have to fold it to AST, and use the AST_unparser). Fixing this in the MIR_unparser is tricky, so its best to fold these statements into the AST within Goto_uppering. This doesnt leave much for the later MIR_to_AST call to do, but thats not a problem.

While debugging this, I found that --debug didnt work as planned. If the pass_manager was used during compilation (say, by the Parse_buffer), then the debug information was reset. This fixes this, and renames the DUMP parameter used throughout Pass_manager to reflect that its really used to indicate whether we're in the main bit (ie the compilation), as opposed to any of the smaller uses of the pass_manager.

This also fixes a bug in the Parse_buffer, where Strings would be unparsed as their pointer, not their value.


This leave a bug that we don't test for, which is that the unparsers dont share state, and so indentation is broken. I'll fix this soon.


git-svn-id: http://phc.googlecode.com/svn/trunk@1258 bbcf20a9-8526-0410-90e2-9b3d3f589b1d