This is a very large change. Ir replaces Assignment with Assign_var, Assign_var_var, Assign_array, Push_array and Invoke_expr. Some slight modification from the original plan was required, as targets weren't handled, so each of these statements now has a Target field. I will remove this from a few of them on a future commit.
Invoke_expr is that same as Assign_var, except with no LHS. Only a number of expression are evocable, namely Method_invocation, Pre_op and New. These are put in a separate Expr called Evocaable_expr.

All other changes follow this, and were mostly done mechanically. A few fixes were needed for statements which werent lowered to the correct form already, but these have been committed in previous patches.

Generate_C has not been fixed (though it compiles), and this breaks about half the tests. This will need larger changes, so I'm OK with that for now. It will be done in the near future.


git-svn-id: http://phc.googlecode.com/svn/trunk@1296 bbcf20a9-8526-0410-90e2-9b3d3f589b1d