Now that the MIR supports Eval_expr, there's no point in converting all statements into assignments. A small number of statements get turned into HIR::Eval_expr, and all the rest get removed. Only Pre_op, Method_invocation and New need to be supported. All single variable statements get removed by a new Remove_solo_variables pass, instead of AST_shredder.
The largest implication here is during codegen. A new pattern, Pattern_eval_expr_or_assign_var deals with the difference between the two possible ways to represent a Method_invocation. The only other statement which needs to be treated this way in New, but we don't generate code for New yet.


git-svn-id: http://phc.googlecode.com/svn/trunk@1316 bbcf20a9-8526-0410-90e2-9b3d3f589b1d