Change Variable_name to to Variable_variable instead of Reflection. This is the first of a number of steps to remove Reflection, and add Variable_name to Expr.
Since the folds were not able to compile with this change before, I added a None type using maketea. Now we return NULL, and save the result for the caller. Since there are a number of callers that use this result, a number of fold_* methods were over-ridden to make isolate the changes. This worked quite well, and I'll use this again elsewhere if I can.


git-svn-id: http://phc.googlecode.com/svn/trunk@1438 bbcf20a9-8526-0410-90e2-9b3d3f589b1d