Remove Variable from the HIR and MIR, and allow Variable_name as the RHS of Assign_var. This is the last of the large changes that we planned when the conversion to HIR/MIR started. There are a lot of changes here, so I'll go into some detail.
Generate_C is for the most part a lot cleaner as a result of this change. Functions were were used to extract VARIABLE_NAMEs from Variables, or wrap one or the other, are now redundant, and are replaced with casts.

Copy_propagation, Dead_code_elimination and Use_def_counter no longer need simple_var, as its now implicit in VARIABLE_NAME.

AST_to_HIR no longer needs is_wrapped_variable_name and expr_to_var_name. Instead, AST::Variables are lowered to HIR::VARIABLE_NAMEs, which can be used directly, or with casts.
Attributes are now copied from a number of places, instead of just one. For example, a VARIABLE_NAME takes the attributes from the enclosing Variable, as well as the VARIABLE_NAME.

The {M,H}IR_unparsers need to handle VARIABLE_NAMEs and Variable_variables explicitly, since the fold for those nodes now returns NULL (see below).

By far the largest changes are to the {H/M}IR_to_AST folders. Since a VARIABLE_NAME is now both an HIR::Expr and a HIR::Variable_name, we must use None as its folded type to satisfy the type-checking. As a result, we do not want to return a None, so we store the result in a class variable. Ideally, we would return NULL then, but NULL is already used by being an optional target. Instead we return 'non_null_ptr', which indicates that the VARIABLE_NAME should come from the class variable. Unlike other class variables, such as reflection and foreign_statement, there can be multiple VARIABLE_NAMEs in use at the same time, so we employ a stack for them, and fetch them in the opposite order. We also abstract the use of None in the manner in which we did for Reflection, because we can't can a None to a VARIABLE_NAME, and we must us None in the signatures of the functions which use VARIABLE_NAMEs.



git-svn-id: http://phc.googlecode.com/svn/trunk@1443 bbcf20a9-8526-0410-90e2-9b3d3f589b1d