Fix a lot of the XML_parser.
Rather than having 3 parsers, there is now one parser, which delegates to handlers based on namespace. This means we can support FOREIGN nodes, which contain nodes from different namespaces. FOREIGN needs seperate work, since it was assumed that a value was a string, but it can now contain nodes as well.

Allow debugging from the start of the parsing, not just during real passes.


git-svn-id: http://phc.googlecode.com/svn/trunk@1520 bbcf20a9-8526-0410-90e2-9b3d3f589b1d