It wasn't possible to reduce test cases at a finer granularity that AST-to-HIR. This allows them to be reduced at the MIR level.
The -p option to reduce now allows a choice of pass to dump on.

--run plugin_name doesnt work with libtool and gdb, whereas the --run=plugin_name form does, so switch to the latter.

A lot of calls ignored the exit_code and stderr; replace these with run_safe, which asserts the result is OK.

Fix plugins to only run once (but that that once is immediately after XML reading.

Minor cleanups to pass_manager.



git-svn-id: http://phc.googlecode.com/svn/trunk@1596 bbcf20a9-8526-0410-90e2-9b3d3f589b1d