Reading and writing to the 'null variable variable' is now supported (see two example tests).
To get this working, it was convenient to have access to a zval* corresponding
to the empty string within the runtime. I've added init_runtime() and
finalize_runtime() functions which are invoked by the phc generated code which
enable to the runtime to initialize its own memory. At the moment this is used
only to initialize the (static) zval "empty_string", but will probably be
useful for more stuff in the future. We might even envision initializing PHP
from within the phc runtime, so that more code is moved to support_routines
from Generate_C.


git-svn-id: http://phc.googlecode.com/svn/trunk@1723 bbcf20a9-8526-0410-90e2-9b3d3f589b1d