Get garbage collection nearly working.
Add Map, Stack and Vector, which are identical to their STL equivalents, except they use the gc_allocator if GC is enabled. This is abstracted by the use of phc_allocator, which is gc_allocator with --enable-gc, and std::allocator otherwise.

Use List::allocator as the default allocator in folds. Without this change (from maketea, revision 137), the signatures of methods with had List parameters was different: the AST_fold used std::allocator, the AST_to_HIR used gc_allocator.

The problems with the GC are much more sporatic now. There remains that each class we define should inherit from GC_obj, and a few edge cases like the parser and Xerces.


git-svn-id: http://phc.googlecode.com/svn/trunk@1790 bbcf20a9-8526-0410-90e2-9b3d3f589b1d