Even though we allow Rvalues in many places, we still cannot generate code for them. This adds a test to skip the shredding stage of most places that allow rvalues, in order to test generating code on them.
A previous experiment which tried to remove shredding of literals deemed it to be too difficult to determine when to shred and when not to. However, allowing literals through the later stages of the pipeline is fairly trivial, so thats probably the easiest way to go.

In related news, the HIR representation of Variable_actual_parameter is probably here to stay, as we dont have another way of representing (Target? Variable_name? Rvalue*). It may be an idea to refactor it to use the same representation in isset and unset, but thats orthogonal to this change.


git-svn-id: http://phc.googlecode.com/svn/trunk@1805 bbcf20a9-8526-0410-90e2-9b3d3f589b1d