I spent a long time trying to reduce the memory used by the refcounts tests for long-tests. The problem is that some tests return 30MB of data. That's copied a few times in homogenize_* functions, then each preg_replace allocates 2x the data required. Then the 3rd part of the test has the same data, so goes through the same process. A test that uses 30MB of data can easily peak over 300MB. After trying to reduce the memory, it seems that the easiest thing is just not to run refcount tests when there is a lot of iteration.
This also adds some steps to reduce the amount of memory. It generally removes one copy of the data, (2 in the case of timeouts).


It also adds a get_num_procs() method, which can be overriden by subclasses. The intension was the the refcounts test would just run 1 process at a time. However, even this is too much in a lot of cases.


git-svn-id: http://phc.googlecode.com/svn/trunk@1858 bbcf20a9-8526-0410-90e2-9b3d3f589b1d