Nearly all uses to is_X_new go through write_var. So make a separate version which doesnt use it, which  we call write_var. The old version is renamed to write_var_TODO. All calls are changed to call write_var_TODO, so that I can change them back one-by-one.
The new write_var uses sep_change_on_write instead of is_X_new. It is the only caller to sep_change_on_write, so remove the 'in_change_on_write' check in sep_change_on_write. In the case of write_var, it must be in the change on write set, or else is_ref would not be set.


git-svn-id: http://phc.googlecode.com/svn/trunk@2014 bbcf20a9-8526-0410-90e2-9b3d3f589b1d