Fix XML_roundtrip.
Add support for nodelist to the XML parser.
We did not support attributes nested within other attributes (say, if a node was an attribute and it the node had attributes as well). The key of the inner attribute was used for the outer attribute. This adds a key stack to fix it.


git-svn-id: http://phc.googlecode.com/svn/trunk@2025 bbcf20a9-8526-0410-90e2-9b3d3f589b1d