Split Generate_c into code for spitting out C, and code for generating LIR. This means that Generate_LIR isnt a pass, but is used like a Fold.
This adds the same framework for the LIR as existed for the other IRs - XML etc.

It was difficult to debug this, since all of the UNINTERPRETED code ended up in base64. This adds a --no-base-64 flag to spit it out as-is (perhaps it should just put it in a comment instead though?).


git-svn-id: http://phc.googlecode.com/svn/trunk@2098 bbcf20a9-8526-0410-90e2-9b3d3f589b1d