Unfortunately this confuses a lot of what could/should be separate commits, but they are difficult to separate after the fact. The main idea is to move method_invocations into macros, but a lot of plumbing was required for that.
The most important change is to allow lists to be passed to macros. If the type of the macro is a list, then nothing changes. However, if the type is to a list, there is instead one macro call for each element of the list (with the other parameters unchanged). This allows use to pass a list of Rvalues to the function_call macro, and allows function_call to instantiate a macro for Rvalue passed to it.

We also allow NULLs to be passed to macros, though the macro shouldnt actually use them. This is useful for allowing multiple versions of a macro, some of which dont use a parameter.

Allow interpolation with the syntax: ${PARAM_NAME}. Otherwise, an interpolation in the string "$NAME_" things the parameter is called NAME_. Spirit does not allow me to specify that the final character should not be an underscore, sadly.

All macros which could use filename, line and column numbers now support have them.

I had 8 'instantiate' functions, taking 1-8 parameters. I reasoned that 8 was enough for anything, and it wasnt too painful to have them, but difficult to avoid. I now need 9 and 10 parameters, and its getting out of hand. I replace them instead with variadic macros and the C++ foreach loop. Variadic macros are supposed to be C99 only, but GCC supports them in C++ as well. Generate_C now uses the INST macro instead of "buf << gen->micg.instantiate (...)".

Finally, fix a small bug with dynamic class defintions, where there was a syntax error in the generated code.




git-svn-id: http://phc.googlecode.com/svn/trunk@2196 bbcf20a9-8526-0410-90e2-9b3d3f589b1d